requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
idna==3.10
mailjet==1.4.1
mailjet-rest==1.3.4
pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
stripe==11.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
